feat(脚本): 添加5.Reality管理

+++ install.sh
@@ -3620,7 +3620,7 @@ EOF
 	]
 }
 EOF
-    else
+    elif [[ -z "$3" ]]; then
         rm /etc/v2ray-agent/xray/conf/04_trojan_TCP_inbounds.json >/dev/null 2>&1
     fi
 
@@ -3651,7 +3651,7 @@ EOF
 ]
 }
 EOF
-    else
+    elif [[ -z "$3" ]]; then
         rm /etc/v2ray-agent/xray/conf/03_VLESS_WS_inbounds.json >/dev/null 2>&1
     fi
 
@@ -3686,7 +3686,7 @@ EOF
     ]
 }
 EOF
-    else
+    elif [[ -z "$3" ]]; then
         rm /etc/v2ray-agent/xray/conf/04_trojan_gRPC_inbounds.json >/dev/null 2>&1
     fi
 
@@ -3716,7 +3716,7 @@ EOF
 ]
 }
 EOF
-    else
+    elif [[ -z "$3" ]]; then
         rm /etc/v2ray-agent/xray/conf/05_VMess_WS_inbounds.json >/dev/null 2>&1
     fi
 
@@ -3743,7 +3743,7 @@ EOF
 ]
 }
 EOF
-    else
+    elif [[ -z "$3" ]]; then
         rm /etc/v2ray-agent/xray/conf/06_VLESS_gRPC_inbounds.json >/dev/null 2>&1
     fi
     # VLESS Vision
@@ -3783,7 +3783,7 @@ EOF
     ]
 }
 EOF
-    else
+    elif [[ -z "$3" ]]; then
         rm /etc/v2ray-agent/xray/conf/02_VLESS_TCP_inbounds.json >/dev/null 2>&1
     fi
 
@@ -3863,7 +3863,7 @@ EOF
 }
 EOF
 
-    else
+    elif [[ -z "$3" ]]; then
         rm /etc/v2ray-agent/xray/conf/07_VLESS_vision_reality_inbounds.json >/dev/null 2>&1
         rm /etc/v2ray-agent/xray/conf/08_VLESS_vision_gRPC_inbounds.json >/dev/null 2>&1
     fi
@@ -3967,7 +3967,7 @@ initSingBoxConfig() {
     ]
 }
 EOF
-    else
+    elif [[ -z "$3" ]]; then
         rm /etc/v2ray-agent/sing-box/conf/config/02_VLESS_TCP_inbounds.json >/dev/null 2>&1
     fi
 
@@ -4010,7 +4010,7 @@ EOF
   ]
 }
 EOF
-    else
+    elif [[ -z "$3" ]]; then
         rm /etc/v2ray-agent/sing-box/conf/config/07_VLESS_vision_reality_inbounds.json >/dev/null 2>&1
     fi
 
@@ -4055,7 +4055,7 @@ EOF
   ]
 }
 EOF
-    else
+    elif [[ -z "$3" ]]; then
         rm /etc/v2ray-agent/sing-box/conf/config/08_VLESS_vision_gRPC_inbounds.json >/dev/null 2>&1
     fi
 
@@ -4089,7 +4089,7 @@ EOF
     ]
 }
 EOF
-    else
+    elif [[ -z "$3" ]]; then
         rm /etc/v2ray-agent/sing-box/conf/config/06_hysteria2_inbounds.json >/dev/null 2>&1
     fi
     if echo "${selectCustomInstallType}" | grep -q 9 || [[ "$1" == "all" ]]; then
@@ -4122,7 +4122,7 @@ EOF
     ]
 }
 EOF
-    else
+    elif [[ -z "$3" ]]; then
         rm /etc/v2ray-agent/sing-box/conf/config/09_tuic_inbounds.json >/dev/null 2>&1
     fi
 }
@@ -7210,8 +7210,17 @@ installSubscribe() {
         mapfile -t result < <(initSingBoxPort "${subscribePort}")
         echoContent yellow " ---> 开始配置订阅的伪装站点\n"
         nginxBlog
+        local httpSubscribeStatus=
 
-        if ( ([[ -n "${selectCustomInstallType}" ]] && ! echo "${selectCustomInstallType}" | grep -q -E "0|1|2|3|4|5|6|9") || ([[ -n "${currentInstallProtocolType}" ]] && ! echo "${currentInstallProtocolType}" | grep -q -E "0|1|2|3|4|5|6|9")) && [[ "${selectInstallType}" == "2" || "${coreInstallType}" == "2" ]]; then
+        if [[ "${selectInstallType}" == "2" || "${coreInstallType}" == "2" ]]; then
+            if [[ -n "${selectCustomInstallType}" ]] && ! echo "${selectCustomInstallType}" | grep -q -E "0|1|2|3|4|5|6|9"; then
+                httpSubscribeStatus=true
+            elif [[ -n "${currentInstallProtocolType}" ]] && ! echo "${currentInstallProtocolType}" | grep -q -E "0|1|2|3|4|5|6|9"; then
+                httpSubscribeStatus=true
+            fi
+        fi
+
+        if [[ "${httpSubscribeStatus}" == "true" ]]; then
             echoContent green "未发现tls证书，使用无加密订阅，可能被运营商拦截。请注意风险"
             read -r -p "是否使用[y/n]？" addNginxSubscribeStatus
             if [[ "${addNginxSubscribeStatus}" != "y" ]]; then
@@ -8084,38 +8093,30 @@ xrayCoreRealityInstall() {
 }
 
 # reality管理
+
 manageReality() {
-    if [[ "${coreInstallType}" == "2" ]]; then
-        echoContent red "\n ---> 此功能仅支持Xray-core内核，请全部安装或者个性化安装Hysteria2"
+    readInstallProtocolType
+    if ! echo "${currentInstallProtocolType}" | grep -q -E "7|8" || [[ -z "${coreInstallType}" ]]; then
+        echoContent red "\n ---> 请先安装Reality协议"
         exit 0
     fi
-    echoContent skyBlue "\n进度  1/1 : reality管理"
-    echoContent red "\n=============================================================="
 
-    if [[ -n "${realityStatus}" ]]; then
-        echoContent yellow "1.重新安装"
-        echoContent yellow "2.卸载"
-        echoContent yellow "3.更换配置"
-    else
-        echoContent yellow "1.安装"
+    if [[ "${coreInstallType}" == "1" ]]; then
+        selectCustomInstallType=7
+        initXrayConfig custom 1 true
+    elif [[ "${coreInstallType}" == "2" ]]; then
+        if echo "${currentInstallProtocolType}" | grep -q "7"; then
+            selectCustomInstallType=7
+        fi
+        if echo "${currentInstallProtocolType}" | grep -q "8"; then
+            selectCustomInstallType=8
+        fi
+        initSingBoxConfig custom 1 true
     fi
-    echoContent yellow "4.扫描Reality域名"
-    echoContent red "=============================================================="
-    read -r -p "请选择:" installRealityStatus
 
-    if [[ "${installRealityStatus}" == "1" ]]; then
-        selectCustomInstallType="7"
-        xrayCoreRealityInstall
-    elif [[ "${installRealityStatus}" == "2" ]]; then
-        unInstallXrayCoreReality
-    elif [[ "${installRealityStatus}" == "3" ]]; then
-        initXrayRealityConfig 1
-        updateXrayRealityConfig
-    elif [[ "${installRealityStatus}" == "4" ]]; then
-        installRealityScanner
-        realityScanner
-    fi
+    reloadCore
 }
+
 # 安装reality scanner
 installRealityScanner() {
     if [[ ! -f "/etc/v2ray-agent/xray/reality_scan/RealiTLScanner-linux-64" ]]; then
@@ -8311,7 +8312,7 @@ menu() {
 	echoContent red "\n=============================================================="
 	echoContent green "原作者：mack-a"
 	echoContent green "作者：Wizard89"
-	echoContent green "当前版本：v2.9.14"
+	echoContent green "当前版本：v2.9.15"
 	echoContent green "Github：https://github.com/Wizard89/v2ray-agent"
 	echoContent green "描述：八合一共存脚本\c"
 	showInstallStatus
@@ -8326,8 +8327,10 @@ menu() {
 	echoContent yellow "2.任意组合安装"
     if [[ "${coreInstallType}" != "2" ]]; then
         echoContent yellow "4.Hysteria2管理"
-        #    echoContent yellow "5.REALITY管理"
+        echoContent yellow "5.REALITY管理"
         echoContent yellow "6.Tuic管理"
+    else
+        echoContent yellow "5.REALITY管理"
     fi
 
 	echoContent skyBlue "-------------------------工具管理-----------------------------"
@@ -8363,9 +8366,9 @@ menu() {
 	4)
 		manageHysteria
 		;;
-        #    5)
-        #        manageReality 1
-        #        ;;
+    5)
+        manageReality 1
+        ;;
     6)
         manageTuic
         ;;
